build fixes. fix strangely broken code
also STATIC_ASSERT is broken. disabled in timer.hpp

Also changed nu_hz's type back to uint32_t to fix lots of warnings.
There are several HZ-like variables in use (inconsistently). these
should be cleaned up / brought into uniform usage (hint)

+++ libs/nupp/param.cpp
@@ -2,4 +2,4 @@
 
 using namespace nu;
 
-uint64_t Param::nu_hz = NU_DEFAULT_HZ;
+uint32_t Param::nu_hz = NU_DEFAULT_HZ;
